feat: enhance mobile navigation and update PageHeader component
- Updated PageHeader to include a mobile navigation toggle button, improving accessibility on mobile devices.
- Refactored CSS for the mobile nav toggle to ensure consistent styling and responsiveness.
- Removed deprecated mobile navigation styles from NavigationWrapper to streamline the codebase.
- Introduced state management for mobile navigation visibility, enhancing user experience.
